+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+
+- Manage errors and exceptions at end of trial
+- Debug logging at critical points
+
+### Fixed
+
+- Bug fix where environment "Final" flag was not being set
+
 ## v1.0.0-alpha4 - 2021-01-26
 
 ### Added